Fix logout.
engine.logout() is slow and authLoggedOut signal is triggered after a while, so we cannot immediately login again because our data will be deleted or partially corrupted (phoneNumber is deleted from profiles.db by TelegramProfileManagerModelItem on authLoggedAuth signal).

Engine state doesn't change in tryInit() call after logout (it's changed to AuthUnknown but it was already set to same value by logout function) and checkAuth() is not called, so AuthPage is not showed; workaround this issue setting engine.phoneNumber = "" on authLoggedOut signal.

Depends on ubports-fixes-old3.x TelegramQML branch.